refactor(shared): hoist sanitizeLogMessage to shared utils
